DT project/requirements.txt
absl-py==2.4.0
asgiref==3.11.1
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==6.0.2
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.61.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.32
modelbenchx==1.0.3
ngrok==1.7.0
nltk==3.9.2
numpy==2.4.2
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.90
packaging==26.0
pandas==3.0.0
pillow==12.1.0
psutil==7.2.2
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
regex==2026.1.15
requests==2.32.5
ruby==0.0.2
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
six==1.17.0
sounddevice==0.5.5
soupsieve==2.8.3
sqlparse==0.5.5
threadpoolctl==3.6.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
